Allow logout to root instead of referrer URL
This commit changes the redirect URL after logout. The original logout
redirect URL was referrer URL. This causes user being redirected back to
login page, which is an unwanted behaviour. Now the user is redirect to
root "/" of the site.

Todo: using logout.target attribute in config.yml file instead of "/"